split tests for playwright and jest
